chore(deps): move to citty 0.2 and cover argument parsing
citty 0.2 is ESM-only, sheds its runtime dependencies, and rewrites the
parser on top of node:util.parseArgs. That last part is what made this a
real bump rather than a number: the release notes call out a change to how
negative boolean arguments print, and `--no-install` / `--no-git` are the
CLI's own negation flags.

Compared both versions on the same definitions across fifteen cases —
negation, aliases, positionals, `--flag=value`, boolean defaults — and the
output is identical. Nothing to absorb.

The command tree moves out of bin.ts into cli.ts, leaving the bin as the
launch only. This is what makes the parsing testable: a module that calls
runMain on import runs the CLI, so nothing could reach the definitions, and
that is precisely why this area had no coverage. Fifteen cases now parse
real argv against the real definitions.

Had the rewrite changed negation, the scaffolder would have installed
dependencies for someone who asked it not to, and nothing downstream would
have complained.
