+++ pnpm-workspace.yaml
@@ -11,7 +11,7 @@ catalog:
   '@types/archiver': ^8.0.0
   '@types/node': ^26.4.0
   archiver: ^8.0.0
-  citty: ^0.1.6
+  citty: ^0.2.2
   happy-dom: ^20.12.0
   knip: ^6.34.0
   lefthook: ^2.1.12
